feat: Add comments relationship to User model and Comment model
This commit adds the comments relationship to the User model and Comment model. The User model now has a hasMany relationship with the Comment model, allowing a user to have multiple comments. The Comment model now belongs to a user, establishing the relationship between users and comments. This change enables the retrieval of a user's comments and the association of comments with their respective users.