feat: Add dark mode support to the application
This commit adds support for dark mode to the application by setting the `class` attribute of the `<html>` element to "dark". This change allows the application to switch between light and dark mode based on the user's preference. The `tailwind.config.js` file was modified to include the `darkMode` property with the value set to "class". Additionally, the `app.blade.php` file was updated to add the `class` attribute with the value "dark" to the `<html>` element.